# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "AI 智能体桥接器 (ai-bridge) API",
    "description": "Trae IDE / 微信 ClawBot 与 AI 智能体之间的通用中间层。",
    "version": "5.5.5",
    "contact": {
      "name": "ai-bridge",
      "url": "https://gitee.com/yzj1/ai-bridge"
    }
  },
  "servers": [
    {
      "url": "http://localhost:4567",
      "description": "本地开发服务器"
    }
  ],
  "security": [
    {
      "cookieAuth": []
    }
  ],
  "paths": {
...
    "/api/auth/login": {
      "post": {
        "summary": "用户登录",
        "tags": ["Auth"],
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": { "type": "string" },
                  "password": { "type": "string" }
                },
                "required": ["username", "password"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "登录成功" },
          "401": { "description": "用户名或密码错误" },
          "429": { "description": "请求过于频繁" }
        }
      }
    },
...
    "/api/auth/me": {
      "get": {
        "summary": "当前登录用户信息",
        "tags": ["Auth"],
        "responses": {
          "200": { "description": "用户信息" },
          "401": { "description": "未登录" }
        }
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "aibridge_access_token",
        "description": "httpOnly cookie 认证；非安全请求需同时提供 X-CSRF-Token header"
      }
    }
  }
}